refactor: add collections resolver
This refactor adds an extendable method responsible for resolving the
collections an entity might be stored in. It does so by accepting kwargs
and returns `["entities"]` by default, as this is the standard for many
clients.

The reason why it returns a list of collections is to have the ability
to have an "autosearch" feature. PZA has a relevant use case for this:
- if you route through the elody spec, the collection name is within
  your path, so this is passed as an argument to the resolve function
  which just returns the same collection name as a list instead of a
  string
- if you go through the ngsi-ld spec you are obligated to use the <urn>
  in the route as id, which includes the entity type and is used to
  determine the collection in the overriden version of the collections
  resolver in PZA
- if you go through the ngsi-ld spec and use an alternate id instead of
  the <urn>, the collections resolver will return all possible
  collections the entity might be stored in, because in this case it
  will not be able to determine the exact collection name

=> in most of the cases you will need to return a list with 1 collection
name, but it being a list has it use cases as described above

Refs #126367